Allow custom scheduler names in `--dist` command line argument
There seems to be no way to fix the argument validation, at least without
complicating things unnecessarily.

So simply remove it. Instead, check `pytest_xdist_make_scheduler` return
value.

Also, convert every builtin scheduler to a separate plugin.

Fixes: https://github.com/pytest-dev/pytest-xdist/issues/970
