+++ changelog/970.feature
@@ -1,0 +1,1 @@
+`--dist` option allows custom scheduler names now.
