+++ src/xdist/scheduler/worksteal.py
@@ -7,6 +7,7 @@ import pytest
 
 from xdist.remote import Producer
 from xdist.report import report_collection_diff
+from xdist.scheduler.protocol import Scheduling
 from xdist.workermanage import parse_spec_config
 from xdist.workermanage import WorkerController
 
@@ -343,3 +344,14 @@ class WorkStealingScheduling:
                     self.config.hook.pytest_collectreport(report=rep)
 
         return same_collection
+
+
+@pytest.hookimpl(trylast=True)
+def pytest_xdist_make_scheduler(
+    config: pytest.Config,
+    log: Producer,
+) -> Scheduling | None:
+    if config.getoption("dist") == "worksteal":
+        return WorkStealingScheduling(config, log)
+    else:
+        return None
